Fix the date so timeslice will work. Use Score.
The timeslice query needs a number timestamp or a date string.
We have a timestamp string. Bummer. For now fix it when adding docs.
TODO: this means adding JSON in the Disco UI won't work.

The Score shown in the UI is always 1.0000. This is boring so use
the score from the review data for now.